Moved ICON CMORizer to designated project 'ICON'
